Fix possible loss of data warning in RGBController_NZXTHuePlus.cpp

+++ Controllers/NZXTHuePlusController/RGBController_NZXTHuePlus.cpp
@@ -302,7 +302,7 @@ void RGBController_HuePlus::DeviceUpdateMode()
                     modes[active_mode].speed,
                     direction,
                     colors,
-                    modes[active_mode].colors.size()
+                    (unsigned int)modes[active_mode].colors.size()
                     );
         }
     }
